Use KNN graph to get Mapper clusters instead of DBSCAN
Since we are already computing the KNN graph, we don't need to introduce an additional dependency on DBSCAN clustering. This also reduces the number of parameters and makes the output of the Mapper algorithm more robust to parameter changes (since the clusters are defined by the KNN graph, which contains global information).